This adds bugfixes, an android API, doc changes
Corrected a problem with connecting on some Android phones running
Android 6/7.

Add a new feature to BGXpressService for writing binary serial data
to a BGX on Android.

Correct a bug in BGXCommander for Android that could cause a
connection to the wrong BGX when selecting a device from the device
list after a previous connnection.

Tag android log messages with "bgx_dbg" for easier message filtering.

Update gradle tool versions to newer versions.

Remove device UUID from Device list in BGXCommander on iOS because
it does not come directly from the device (generated by Core Bluetooth)
and is confusing.

Fix an iOS bug that prevented the bus mode from being changed from
LOCAL_COMMAND_MODE.

Add a note about the need to install CocoaPods in order to build
BGXCommander for iOS to the README.md.